% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateEigengene}
\alias{calculateEigengene}
\title{Calculate eigengene}
\usage{
calculateEigengene(matrix)
}
\arguments{
\item{matrix}{A matrix for which the eigengene is to be calculated.}
}
\value{
The eigengene of the input matrix.
}
\description{
Calculates the eigengene for a given matrix. We use the \emph{irlba} package to efficiently calculate the eigengene using the Lanzcos algorithm. The eigengene is aligned to match the orientation of the input matrix and will retain the row names of the input matrix.
}
\details{
The eigengene is canonically known as the first principal component. However, principal components are rotation-invariant, meaning that the result can be sometimes inverted (\emph{i.e. the results times -1}). We orient the eigengene to the input matrix by aligning the correlation between the eigengene and the vector of averages (\emph{row means}) of the input matrix.
}
\examples{
# Generate data
obs <- rbind(
  sapply(1:10, function(x) {
    rnorm(100, mean = x)
  }),
  sapply(10:1, function(x) {
    rnorm(100, mean = x)
  })
)

# Calculate Eigengene
eg <- calculateEigengene(obs)

# Eigengene should reflect mean differences
plot(eg, ylab = "Eigengene Value")
}
\seealso{
\code{\link[irlba:irlba]{irlba::irlba()}}
}
