% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CILP}
\alias{CILP}
\title{Run Correlation by Individual Level Product (CILP)}
\usage{
CILP(prod_mat, groupings)
}
\arguments{
\item{prod_mat}{A matrix with all class-specific mean-removed pairwise interactions to be tested.}

\item{groupings}{A numeric vector denoting class/group membership of the observations.}
}
\value{
A matrix of effect size estimates and p-values for each feature pair.
}
\description{
Run Correlation by Individual Level Product (CILP)
}
