% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{demean}
\alias{demean}
\title{Remove class-specific mean from each class in matrix}
\usage{
demean(mat, class_df)
}
\arguments{
\item{mat}{A matrix with observations (e.g. individuals) as rows and features as columns.}

\item{class_df}{A data frame with two columns. The first columns are observation identifiers and the second column denotes class/group.}
}
\value{
A matrix with the class-specific mean removed from each observation in its appropriate group.
}
\description{
Remove class-specific mean from each class in matrix
}
