% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_interactions}
\alias{generate_interactions}
\title{Generate pairwise interaction matrix}
\usage{
generate_interactions(matrix, delimiter = "_", interactions_df = NA)
}
\arguments{
\item{matrix}{A matrix with observations (e.g. individuals) as rows and
features as columns.}

\item{delimiter}{A delimiter for the pairwise feature names.}

\item{interactions_df}{A two-column data frame of interactions to be
calculated. The first column denotes the first feature and the second
column denotes the second feature of the interaction.}
}
\value{
A matrix with all pairwise interactions (pairwise column/feature
products) from the input matrix. Column names will show what features were
used for the pairwise product.
}
\description{
This function generates a matrix where the columns/features are the pairwise
products of the input matrix features. The new matrix will retain the same
number of rows/observations, but have \emph{m} choose 2 columns for each pair
of features, where \emph{m} is the number of features in the input matrix.
When using this function where the mean has been removed, the feature/column
means become observations of covariance.
}
\details{
The pairwise interaction matrix allows one to estimate covariance when paired with the demean function. This relies on the fact that \verb{Cov(X,Y) = E[XY]-E[X]E[Y]}, but by removing means, we eliminate the second term.
}
\examples{
# Generate covariance matrix
library(mvtnorm)

cov_mat <- rWishart(1, 10, diag(10))[, , 1]

# Generate correlated data with means 1-10
dat <- rmvnorm(100, mean = 1:10, sigma = cov_mat)
colnames(dat) <- LETTERS[1:10]
rownames(dat) <- 1:100

# Remove mean from matrix
centered_dat <- scale(dat, scale = FALSE)
colMeans(centered_dat)

# Generate pairwise matrix
product_mat <- generate_interactions(centered_dat)

# Check if covariance estimates are close
demean_est <- colSums(product_mat)/(nrow(product_mat)-1)

samp_cov <- sapply(strsplit(colnames(product_mat),split="_"),function(x){cov(dat)[x[1],x[2]]})
names(samp_cov) <- colnames(product_mat)

head(demean_est)
head(samp_cov)
}
\seealso{
\code{\link[stats:cor]{stats::cov()}}
}
