% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_squared_terms}
\alias{generate_squared_terms}
\title{Generate squared matrix}
\usage{
generate_squared_terms(mat)
}
\arguments{
\item{mat}{A mean-removed matrix with observations (e.g. individuals) as rows and features as columns.}
}
\value{
A matrix with each column/feature squared.
}
\description{
Generate squared matrix
}
