% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateEigengene}
\alias{calculateEigengene}
\title{Calculate eigengene}
\usage{
calculateEigengene(mat)
}
\arguments{
\item{mat}{A matrix for which the eigengene is to be calculated.}
}
\value{
The eigengene of the input matrix.
}
\description{
Calculate eigengene
}
\examples{
# Generate data
obs <- rbind(sapply(1:10,function(x){rnorm(100,mean=x)}), 
       sapply(10:1,function(x){rnorm(100,mean=x)}))

# Calculate Eigengene
eg <- calculateEigengene(obs)

# Eigengene should reflect mean differences
plot(eg,ylab='Eigengene Value')
}
