% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateEigengene}
\alias{calculateEigengene}
\title{Calculate eigengene}
\usage{
calculateEigengene(mat)
}
\arguments{
\item{mat}{A matrix for which the eigengene is to be calculated.}
}
\value{
The eigengene of the input matrix.
}
\description{
Calculate eigengene
}
