% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_all_interaction}
\alias{generate_all_interaction}
\title{Generate pairwise interaction matrix}
\usage{
generate_all_interaction(mat)
}
\arguments{
\item{mat}{A mean-removed matrix with observations (e.g. individuals) as rows and features as columns.}
}
\value{
A matrix with all pairwise interactions (pairwise column/feature products) from the input matrix.
}
\description{
Generate pairwise interaction matrix
}
